feat(db): scaling options on creation
